+++ glass-slipper/GlassSlipper-Bridging-Header.h
@@ -1,9 +1,0 @@
-//
-//  GlassSlipper-Bridging-Header.h
-//  GlassSlipper
-//
-//  Exposes ObjC headers to Swift.
-//
-
-#import "AppDelegate.h"
-#import "DiagnosticStepCell.h"